Profile Page is now interchangable between personal user profile and clicked profile from Mentor page. Profile Page takes in data to display proper information instead of using authentication for current user
